Razdvajanje logike klijenta
Skripta akdb_cli.py razdvojena je na dvije skripte, odnosno, klasa Client je
izdvojena u zasebnu datoteku. Kako bi kod i dalje funkcionirao, u konstruktor
klijenta dodana su dva parametra: username i password.
